@@ -1,3 +1,4 @@
+// Package table provides a styled table renderer for terminals.
 package table
 
 import (
@@ -42,6 +43,7 @@ func DefaultStyles(_, _ int) lipgloss.Style {
 
 // Table is a type for rendering tables.
 type Table struct {
+	baseStyle lipgloss.Style
 	styleFunc StyleFunc
 	border    lipgloss.Border
 
@@ -60,29 +62,31 @@ type Table struct {
 	width           int
 	height          int
 	useManualHeight bool
-	offset          int
+	yOffset         int
+	wrap            bool
 
-	// widths tracks the width of each column.
-	widths []int
-
-	// heights tracks the height of each row.
+	widths  []int
 	heights []int
+
+	firstVisibleRowIndex int
+	lastVisibleRowIndex  int
 }
 
 // New returns a new Table that can be modified through different
 // attributes.
 //
-// By default, a table has no border, no styling, and no rows.
+// By default, a table has normal border, no styling, and no rows.
 func New() *Table {
 	return &Table{
 		styleFunc:    DefaultStyles,
-		border:       lipgloss.RoundedBorder(),
+		border:       lipgloss.NormalBorder(),
 		borderBottom: true,
 		borderColumn: true,
 		borderHeader: true,
 		borderLeft:   true,
 		borderRight:  true,
 		borderTop:    true,
+		wrap:         true,
 		data:         NewStringData(),
 	}
 }
@@ -90,6 +94,14 @@ func New() *Table {
 // ClearRows clears the table rows.
 func (t *Table) ClearRows() *Table {
 	t.data = NewStringData()
+	return t
+}
+
+// BaseStyle sets the base style for the whole table. If you need to set a
+// background color for the whole table, use this.
+func (t *Table) BaseStyle(baseStyle lipgloss.Style) *Table {
+	t.baseStyle = baseStyle
+	t.borderStyle = t.borderStyle.Inherit(baseStyle)
 	return t
 }
 
@@ -102,15 +114,20 @@ func (t *Table) StyleFunc(style StyleFunc) *Table {
 // style returns the style for a cell based on it's position (row, column).
 func (t *Table) style(row, col int) lipgloss.Style {
 	if t.styleFunc == nil {
-		return lipgloss.NewStyle()
+		return t.baseStyle
 	}
-	return t.styleFunc(row, col)
+	return t.styleFunc(row, col).Inherit(t.baseStyle)
 }
 
 // Data sets the table data.
 func (t *Table) Data(data Data) *Table {
 	t.data = data
 	return t
+}
+
+// GetData returns the table data.
+func (t *Table) GetData() Data {
+	return t.data
 }
 
 // Rows appends rows to the table data.
@@ -137,6 +154,11 @@ func (t *Table) Row(row ...string) *Table {
 func (t *Table) Headers(headers ...string) *Table {
 	t.headers = headers
 	return t
+}
+
+// GetHeaders returns the table headers.
+func (t *Table) GetHeaders() []string {
+	return t.headers
 }
 
 // Border sets the table border.
@@ -189,8 +211,43 @@ func (t *Table) BorderRow(v bool) *Table {
 
 // BorderStyle sets the style for the table border.
 func (t *Table) BorderStyle(style lipgloss.Style) *Table {
-	t.borderStyle = style
+	t.borderStyle = style.Inherit(t.baseStyle)
 	return t
+}
+
+// GetBorderTop gets the top border.
+func (t *Table) GetBorderTop() bool {
+	return t.borderTop
+}
+
+// GetBorderBottom gets the bottom border.
+func (t *Table) GetBorderBottom() bool {
+	return t.borderBottom
+}
+
+// GetBorderLeft gets the left border.
+func (t *Table) GetBorderLeft() bool {
+	return t.borderLeft
+}
+
+// GetBorderRight gets the right border.
+func (t *Table) GetBorderRight() bool {
+	return t.borderRight
+}
+
+// GetBorderHeader gets the header separator border.
+func (t *Table) GetBorderHeader() bool {
+	return t.borderHeader
+}
+
+// GetBorderColumn gets the column border separator.
+func (t *Table) GetBorderColumn() bool {
+	return t.borderColumn
+}
+
+// GetBorderRow gets the row border separator.
+func (t *Table) GetBorderRow() bool {
+	return t.borderRow
 }
 
 // Width sets the table width, this auto-sizes the columns to fit the width by
@@ -208,9 +265,45 @@ func (t *Table) Height(h int) *Table {
 	return t
 }
 
-// Offset sets the table rendering offset.
-func (t *Table) Offset(o int) *Table {
-	t.offset = o
+// GetHeight returns the height of the table.
+func (t *Table) GetHeight() int {
+	return t.height
+}
+
+// YOffset sets the table rendering offset.
+func (t *Table) YOffset(o int) *Table {
+	t.yOffset = o
+	return t
+}
+
+// GetYOffset returns the table rendering offset.
+func (t *Table) GetYOffset() int {
+	return t.yOffset
+}
+
+// FirstVisibleRowIndex returns the index of the first visible row in the table.
+func (t *Table) FirstVisibleRowIndex() int {
+	return t.firstVisibleRowIndex
+}
+
+// LastVisibleRowIndex returns the index of the last visible row in the table.
+func (t *Table) LastVisibleRowIndex() int {
+	return t.lastVisibleRowIndex
+}
+
+// VisibleRows returns the number of visible rows in the table.
+func (t *Table) VisibleRows() int {
+	if t.lastVisibleRowIndex == -2 {
+		return t.data.Rows() - t.firstVisibleRowIndex
+	}
+	return t.lastVisibleRowIndex - t.firstVisibleRowIndex + 1
+}
+
+// Wrap dictates whether or not the table content should wrap.
+//
+// This only applies to data cells. Headers are never wrapped.
+func (t *Table) Wrap(w bool) *Table {
+	t.wrap = w
 	return t
 }
 
@@ -231,120 +324,8 @@ func (t *Table) String() string {
 		}
 	}
 
-	// Initialize the widths.
-	t.widths = make([]int, max(len(t.headers), t.data.Columns()))
-	t.heights = make([]int, btoi(hasHeaders)+t.data.Rows())
-
-	// The style function may affect width of the table. It's possible to set
-	// the StyleFunc after the headers and rows. Update the widths for a final
-	// time.
-	for i, cell := range t.headers {
-		t.widths[i] = max(t.widths[i], lipgloss.Width(t.style(HeaderRow, i).Render(cell)))
-		t.heights[0] = max(t.heights[0], lipgloss.Height(t.style(HeaderRow, i).Render(cell)))
-	}
-
-	for r := 0; r < t.data.Rows(); r++ {
-		for i := 0; i < t.data.Columns(); i++ {
-			cell := t.data.At(r, i)
-
-			rendered := t.style(r, i).Render(cell)
-			t.heights[r+btoi(hasHeaders)] = max(t.heights[r+btoi(hasHeaders)], lipgloss.Height(rendered))
-			t.widths[i] = max(t.widths[i], lipgloss.Width(rendered))
-		}
-	}
-
-	// Table Resizing Logic.
-	//
-	// Given a user defined table width, we must ensure the table is exactly that
-	// width. This must account for all borders, column, separators, and column
-	// data.
-	//
-	// In the case where the table is narrower than the specified table width,
-	// we simply expand the columns evenly to fit the width.
-	// For example, a table with 3 columns takes up 50 characters total, and the
-	// width specified is 80, we expand each column by 10 characters, adding 30
-	// to the total width.
-	//
-	// In the case where the table is wider than the specified table width, we
-	// _could_ simply shrink the columns evenly but this would result in data
-	// being truncated (perhaps unnecessarily). The naive approach could result
-	// in very poor cropping of the table data. So, instead of shrinking columns
-	// evenly, we calculate the median non-whitespace length of each column, and
-	// shrink the columns based on the largest median.
-	//
-	// For example,
-	//  ┌──────┬───────────────┬──────────┐
-	//  │ Name │ Age of Person │ Location │
-	//  ├──────┼───────────────┼──────────┤
-	//  │ Kini │ 40            │ New York │
-	//  │ Eli  │ 30            │ London   │
-	//  │ Iris │ 20            │ Paris    │
-	//  └──────┴───────────────┴──────────┘
-	//
-	// Median non-whitespace length  vs column width of each column:
-	//
-	// Name: 4 / 5
-	// Age of Person: 2 / 15
-	// Location: 6 / 10
-	//
-	// The biggest difference is 15 - 2, so we can shrink the 2nd column by 13.
-
-	width := t.computeWidth()
-
-	if width < t.width && t.width > 0 {
-		// Table is too narrow, expand the columns evenly until it reaches the
-		// desired width.
-		var i int
-		for width < t.width {
-			t.widths[i]++
-			width++
-			i = (i + 1) % len(t.widths)
-		}
-	} else if width > t.width && t.width > 0 {
-		// Table is too wide, calculate the median non-whitespace length of each
-		// column, and shrink the columns based on the largest difference.
-		columnMedians := make([]int, len(t.widths))
-		for c := range t.widths {
-			trimmedWidth := make([]int, t.data.Rows())
-			for r := 0; r < t.data.Rows(); r++ {
-				renderedCell := t.style(r+btoi(hasHeaders), c).Render(t.data.At(r, c))
-				nonWhitespaceChars := lipgloss.Width(strings.TrimRight(renderedCell, " "))
-				trimmedWidth[r] = nonWhitespaceChars + 1
-			}
-
-			columnMedians[c] = median(trimmedWidth)
-		}
-
-		// Find the biggest differences between the median and the column width.
-		// Shrink the columns based on the largest difference.
-		differences := make([]int, len(t.widths))
-		for i := range t.widths {
-			differences[i] = t.widths[i] - columnMedians[i]
-		}
-
-		for width > t.width {
-			index, _ := largest(differences)
-			if differences[index] < 1 {
-				break
-			}
-
-			shrink := min(differences[index], width-t.width)
-			t.widths[index] -= shrink
-			width -= shrink
-			differences[index] = 0
-		}
-
-		// Table is still too wide, begin shrinking the columns based on the
-		// largest column.
-		for width > t.width {
-			index, _ := largest(t.widths)
-			if t.widths[index] < 1 {
-				break
-			}
-			t.widths[index]--
-			width--
-		}
-	}
+	// Do all the sizing calculations for width and height.
+	t.resize()
 
 	var sb strings.Builder
 
@@ -355,7 +336,6 @@ func (t *Table) String() string {
 
 	if hasHeaders {
 		sb.WriteString(t.constructHeaders())
-		sb.WriteString("\n")
 	}
 
 	var bottom string
@@ -365,41 +345,20 @@ func (t *Table) String() string {
 
 	// If there are no data rows render nothing.
 	if t.data.Rows() > 0 {
-		switch {
-		case t.useManualHeight:
-			// The height of the top border. Subtract 1 for the newline.
-			topHeight := lipgloss.Height(sb.String()) - 1
-			availableLines := t.height - (topHeight + lipgloss.Height(bottom))
-
-			// if the height is larger than the number of rows, use the number
-			// of rows.
-			if availableLines > t.data.Rows() {
-				availableLines = t.data.Rows()
+		for r := t.firstVisibleRowIndex; r < t.data.Rows(); r++ {
+			if t.lastVisibleRowIndex != -2 && r > t.lastVisibleRowIndex {
+				break
 			}
-			sb.WriteString(t.constructRows(availableLines))
-
-		default:
-			for r := t.offset; r < t.data.Rows(); r++ {
-				sb.WriteString(t.constructRow(r, false))
-			}
+			sb.WriteString(t.constructRow(r))
 		}
 	}
 
 	sb.WriteString(bottom)
 
 	return lipgloss.NewStyle().
-		MaxHeight(t.computeHeight()).
+		MaxHeight(min(t.height, t.computeHeight())).
 		MaxWidth(t.width).
-		Render(sb.String())
-}
-
-// computeWidth computes the width of the table in it's current configuration.
-func (t *Table) computeWidth() int {
-	width := sum(t.widths) + btoi(t.borderLeft) + btoi(t.borderRight)
-	if t.borderColumn {
-		width += len(t.widths) - 1
-	}
-	return width
+		Render(strings.TrimSuffix(sb.String(), "\n"))
 }
 
 // computeHeight computes the height of the table in it's current configuration.
@@ -422,7 +381,7 @@ func (t *Table) constructTopBorder() string {
 	if t.borderLeft {
 		s.WriteString(t.borderStyle.Render(t.border.TopLeft))
 	}
-	for i := 0; i < len(t.widths); i++ {
+	for i := range t.widths {
 		s.WriteString(t.borderStyle.Render(strings.Repeat(t.border.Top, t.widths[i])))
 		if i < len(t.widths)-1 && t.borderColumn {
 			s.WriteString(t.borderStyle.Render(t.border.MiddleTop))
@@ -441,7 +400,7 @@ func (t *Table) constructBottomBorder() string {
 	if t.borderLeft {
 		s.WriteString(t.borderStyle.Render(t.border.BottomLeft))
 	}
-	for i := 0; i < len(t.widths); i++ {
+	for i := range t.widths {
 		s.WriteString(t.borderStyle.Render(strings.Repeat(t.border.Bottom, t.widths[i])))
 		if i < len(t.widths)-1 && t.borderColumn {
 			s.WriteString(t.borderStyle.Render(t.border.MiddleBottom))
@@ -457,28 +416,48 @@ func (t *Table) constructBottomBorder() string {
 // header configuration and data.
 func (t *Table) constructHeaders() string {
 	var s strings.Builder
+	cells := make([]string, 0, len(t.headers)*2+1)
+	height := t.heights[0]
+
+	left := strings.Repeat(t.borderStyle.Render(t.border.Left)+"\n", height)
 	if t.borderLeft {
-		s.WriteString(t.borderStyle.Render(t.border.Left))
+		cells = append(cells, left)
 	}
-	for i, header := range t.headers {
-		s.WriteString(t.style(HeaderRow, i).
-			MaxHeight(1).
-			Width(t.widths[i]).
-			MaxWidth(t.widths[i]).
-			Render(ansi.Truncate(header, t.widths[i], "…")))
-		if i < len(t.headers)-1 && t.borderColumn {
-			s.WriteString(t.borderStyle.Render(t.border.Left))
+
+	for j, header := range t.headers {
+		cellStyle := t.style(HeaderRow, j)
+
+		// NOTE(@andreynering): We always truncate headers.
+		header = t.truncateCell(header, HeaderRow, j)
+
+		cells = append(cells,
+			cellStyle.
+				Height(height-cellStyle.GetVerticalMargins()).
+				Width(t.widths[j]-cellStyle.GetHorizontalMargins()).
+				Render(header),
+		)
+
+		if j < len(t.headers)-1 && t.borderColumn {
+			cells = append(cells, left)
 		}
 	}
+
+	if t.borderRight {
+		right := strings.Repeat(t.borderStyle.Render(t.border.Right)+"\n", height)
+		cells = append(cells, right)
+	}
+
+	for i, cell := range cells {
+		cells[i] = strings.TrimRight(cell, "\n")
+	}
+
+	s.WriteString(lipgloss.JoinHorizontal(lipgloss.Top, cells...) + "\n")
+
 	if t.borderHeader {
-		if t.borderRight {
-			s.WriteString(t.borderStyle.Render(t.border.Right))
-		}
-		s.WriteString("\n")
 		if t.borderLeft {
 			s.WriteString(t.borderStyle.Render(t.border.MiddleLeft))
 		}
-		for i := 0; i < len(t.headers); i++ {
+		for i := range t.headers {
 			s.WriteString(t.borderStyle.Render(strings.Repeat(t.border.Top, t.widths[i])))
 			if i < len(t.headers)-1 && t.borderColumn {
 				s.WriteString(t.borderStyle.Render(t.border.Middle))
@@ -487,79 +466,49 @@ func (t *Table) constructHeaders() string {
 		if t.borderRight {
 			s.WriteString(t.borderStyle.Render(t.border.MiddleRight))
 		}
+		s.WriteString("\n")
 	}
-	if t.borderRight && !t.borderHeader {
-		s.WriteString(t.borderStyle.Render(t.border.Right))
-	}
+
 	return s.String()
-}
-
-func (t *Table) constructRows(availableLines int) string {
-	var sb strings.Builder
-
-	// The number of rows to render after removing the offset.
-	offsetRowCount := t.data.Rows() - t.offset
-
-	// The number of rows to render. We always render at least one row.
-	rowsToRender := availableLines
-	rowsToRender = max(rowsToRender, 1)
-
-	// Check if we need to render an overflow row.
-	needsOverflow := rowsToRender < offsetRowCount
-
-	// only use the offset as the starting value if there is overflow.
-	rowIdx := t.offset
-	if !needsOverflow {
-		// if there is no overflow, just render to the height of the table
-		// check there's enough content to fill the table
-		rowIdx = t.data.Rows() - rowsToRender
-	}
-	for rowsToRender > 0 && rowIdx < t.data.Rows() {
-		// Whenever the height is too small to render all rows, the bottom row will be an overflow row (ellipsis).
-		isOverflow := needsOverflow && rowsToRender == 1
-
-		sb.WriteString(t.constructRow(rowIdx, isOverflow))
-
-		rowIdx++
-		rowsToRender--
-	}
-	return sb.String()
 }
 
 // constructRow constructs the row for the table given an index and row data
 // based on the current configuration. If isOverflow is true, the row is
 // rendered as an overflow row (using ellipsis).
-func (t *Table) constructRow(index int, isOverflow bool) string {
+func (t *Table) constructRow(index int) string {
 	var s strings.Builder
+	cells := make([]string, 0, t.data.Columns()*2+1)
 
 	hasHeaders := len(t.headers) > 0
 	height := t.heights[index+btoi(hasHeaders)]
+	isLastRow := index == t.data.Rows()-1
+	isOverflow := !isLastRow && t.lastVisibleRowIndex == index
 	if isOverflow {
-		height = 1
+		height = max(height, 1)
 	}
 
-	var cells []string
 	left := strings.Repeat(t.borderStyle.Render(t.border.Left)+"\n", height)
 	if t.borderLeft {
 		cells = append(cells, left)
 	}
 
-	for c := 0; c < t.data.Columns(); c++ {
-		cellWidth := t.widths[c]
-
+	for c := range t.data.Columns() {
 		cell := "…"
 		if !isOverflow {
 			cell = t.data.At(index, c)
 		}
 
 		cellStyle := t.style(index, c)
+		if !t.wrap {
+			cell = t.truncateCell(cell, index, c)
+		}
 		cells = append(cells, cellStyle.
 			// Account for the margins in the cell sizing.
 			Height(height-cellStyle.GetVerticalMargins()).
 			MaxHeight(height).
 			Width(t.widths[c]-cellStyle.GetHorizontalMargins()).
 			MaxWidth(t.widths[c]).
-			Render(ansi.Truncate(cell, cellWidth*height, "…")))
+			Render(cell))
 
 		if c < t.data.Columns()-1 && t.borderColumn {
 			cells = append(cells, left)
@@ -577,16 +526,36 @@ func (t *Table) constructRow(index int, isOverflow bool) string {
 
 	s.WriteString(lipgloss.JoinHorizontal(lipgloss.Top, cells...) + "\n")
 
-	if t.borderRow && index < t.data.Rows()-1 {
-		s.WriteString(t.borderStyle.Render(t.border.MiddleLeft))
-		for i := 0; i < len(t.widths); i++ {
+	if t.borderRow && !isOverflow && index < t.data.Rows()-1 {
+		if t.borderLeft {
+			s.WriteString(t.borderStyle.Render(t.border.MiddleLeft))
+		}
+		for i := range t.widths {
 			s.WriteString(t.borderStyle.Render(strings.Repeat(t.border.Bottom, t.widths[i])))
 			if i < len(t.widths)-1 && t.borderColumn {
 				s.WriteString(t.borderStyle.Render(t.border.Middle))
 			}
 		}
-		s.WriteString(t.borderStyle.Render(t.border.MiddleRight) + "\n")
+		if t.borderRight {
+			s.WriteString(t.borderStyle.Render(t.border.MiddleRight))
+		}
+		s.WriteString("\n")
 	}
 
 	return s.String()
+}
+
+func (t *Table) truncateCell(cell string, rowIndex, colIndex int) string {
+	hasHeaders := len(t.headers) > 0
+	height := t.heights[rowIndex+btoi(hasHeaders)]
+	cellWidth := t.widths[colIndex]
+	cellStyle := t.style(rowIndex, colIndex)
+
+	// NOTE(@andreynering): We always truncate headers to 1 line.
+	if rowIndex == HeaderRow {
+		height = 1
+	}
+
+	length := (cellWidth * height) - cellStyle.GetHorizontalPadding() - cellStyle.GetHorizontalMargins()
+	return ansi.Truncate(cell, length, "…")
 }
