@@ -1,6 +1,8 @@
 package lipgloss
 
 import (
+	"cmp"
+	"errors"
 	"image/color"
 	"strconv"
 	"strings"
@@ -9,6 +11,13 @@ import (
 	"github.com/charmbracelet/x/ansi"
 	"github.com/lucasb-eyer/go-colorful"
 )
+
+func clamp[T cmp.Ordered](v, low, high T) T {
+	if high < low {
+		high, low = low, high
+	}
+	return min(high, max(low, v))
+}
 
 // 4-bit color constants.
 const (
@@ -57,33 +66,72 @@ func (n NoColor) RGBA() (r, g, b, a uint32) {
 //	ansi256Color := lipgloss.Color("21")
 //	hexColor := lipgloss.Color("#0000ff")
 func Color(s string) color.Color {
-	var c color.Color = noColor
 	if strings.HasPrefix(s, "#") {
-		hex, err := colorful.Hex(s)
+		c, err := parseHex(s)
 		if err != nil {
-			return c
+			return noColor
 		}
-		c = hex
-	} else {
-		i, err := strconv.Atoi(s)
-		if err != nil {
-			return c
-		}
-
-		if i < 0 {
-			// Only positive numbers
-			i = -i
-		}
-
-		if i < 16 {
-			c = ansi.BasicColor(i) //nolint:gosec
-		} else if i < 256 {
-			c = ANSIColor(i) //nolint:gosec
-		} else {
-			c = ansi.TrueColor(i) //nolint:gosec
-		}
+		return c
 	}
-	return c
+
+	i, err := strconv.Atoi(s)
+	if err != nil {
+		return noColor
+	}
+
+	if i < 0 {
+		// Only positive numbers
+		i = -i
+	}
+
+	if i < 16 {
+		return ansi.BasicColor(i) //nolint:gosec
+	} else if i < 256 {
+		return ANSIColor(i) //nolint:gosec
+	}
+
+	r, g, b := uint8((i>>16)&0xff), uint8(i>>8&0xff), uint8(i&0xff) //nolint:gosec
+	return color.RGBA{R: r, G: g, B: b, A: 0xff}
+}
+
+var errInvalidFormat = errors.New("invalid hex format") // pre-allocated.
+
+// parseHex parses a hex color string and returns a color.RGBA. The string can be
+// in the format #RRGGBB or #RGB. This is a more performant implementation of
+// [colorful.Hex].
+func parseHex(s string) (c color.RGBA, err error) {
+	c.A = 0xff
+
+	if len(s) == 0 || s[0] != '#' {
+		return c, errInvalidFormat
+	}
+
+	hexToByte := func(b byte) byte {
+		switch {
+		case b >= '0' && b <= '9':
+			return b - '0'
+		case b >= 'a' && b <= 'f':
+			return b - 'a' + 10
+		case b >= 'A' && b <= 'F':
+			return b - 'A' + 10
+		}
+		err = errInvalidFormat
+		return 0
+	}
+
+	switch len(s) {
+	case 7:
+		c.R = hexToByte(s[1])<<4 + hexToByte(s[2])
+		c.G = hexToByte(s[3])<<4 + hexToByte(s[4])
+		c.B = hexToByte(s[5])<<4 + hexToByte(s[6])
+	case 4:
+		c.R = hexToByte(s[1]) * 17
+		c.G = hexToByte(s[2]) * 17
+		c.B = hexToByte(s[3]) * 17
+	default:
+		err = errInvalidFormat
+	}
+	return c, err
 }
 
 // RGBColor is a color specified by red, green, and blue values.
@@ -110,7 +158,7 @@ func (c RGBColor) RGBA() (r, g, b, a uint32) {
 //
 //	colorA := lipgloss.ANSIColor(8)
 //	colorB := lipgloss.ANSIColor(134)
-type ANSIColor = ansi.ExtendedColor
+type ANSIColor = ansi.IndexedColor
 
 // LightDarkFunc is a function that returns a color based on whether the
 // terminal has a light or dark background. You can create one of these with
@@ -216,7 +264,7 @@ type CompleteFunc func(ansi, ansi256, truecolor color.Color) color.Color
 //	fmt.Println("Ooh, pretty color: ", color)
 func Complete(p colorprofile.Profile) CompleteFunc {
 	return func(ansi, ansi256, truecolor color.Color) color.Color {
-		switch p {
+		switch p { //nolint:exhaustive
 		case colorprofile.ANSI:
 			return ansi
 		case colorprofile.ANSI256:
@@ -225,5 +273,87 @@ func Complete(p colorprofile.Profile) CompleteFunc {
 			return truecolor
 		}
 		return noColor
+	}
+}
+
+// ensureNotTransparent ensures that the alpha value of a color is not 0, and if
+// it is, we will set it to 1. This is useful for when we are converting from
+// RGB -> RGBA, and the alpha value is lost in the conversion for gradient purposes.
+func ensureNotTransparent(c color.Color) color.Color {
+	_, _, _, a := c.RGBA()
+	if a == 0 {
+		return Alpha(c, 1)
+	}
+	return c
+}
+
+// Alpha adjusts the alpha value of a color using a 0-1 (clamped) float scale
+// 0 = transparent, 1 = opaque.
+func Alpha(c color.Color, alpha float64) color.Color {
+	if c == nil {
+		return nil
+	}
+
+	r, g, b, _ := c.RGBA()
+	return color.RGBA{
+		R: uint8(min(255, float64(r>>8))),
+		G: uint8(min(255, float64(g>>8))),
+		B: uint8(min(255, float64(b>>8))),
+		A: uint8(clamp(alpha, 0, 1) * 255),
+	}
+}
+
+// Complementary returns the complementary color (180° away on color wheel) of
+// the given color. This is useful for creating a contrasting color.
+func Complementary(c color.Color) color.Color {
+	if c == nil {
+		return nil
+	}
+
+	// Offset hue by 180°.
+	cf, _ := colorful.MakeColor(ensureNotTransparent(c))
+
+	h, s, v := cf.Hsv()
+	h += 180
+	if h >= 360 {
+		h -= 360
+	} else if h < 0 {
+		h += 360
+	}
+
+	return colorful.Hsv(h, s, v).Clamped()
+}
+
+// Darken takes a color and makes it darker by a specific percentage (0-1, clamped).
+func Darken(c color.Color, percent float64) color.Color {
+	if c == nil {
+		return nil
+	}
+
+	mult := 1.0 - clamp(percent, 0, 1)
+
+	r, g, b, a := c.RGBA()
+	return color.RGBA{
+		R: uint8(float64(r>>8) * mult),
+		G: uint8(float64(g>>8) * mult),
+		B: uint8(float64(b>>8) * mult),
+		A: uint8(min(255, float64(a>>8))),
+	}
+}
+
+// Lighten makes a color lighter by a specific percentage (0-1, clamped).
+func Lighten(c color.Color, percent float64) color.Color {
+	if c == nil {
+		return nil
+	}
+
+	add := 255 * (clamp(percent, 0, 1))
+
+	r, g, b, a := c.RGBA()
+	return color.RGBA{
+		R: uint8(min(255, float64(r>>8)+add)),
+		G: uint8(min(255, float64(g>>8)+add)),
+		B: uint8(min(255, float64(b>>8)+add)),
+		A: uint8(min(255, float64(a>>8))),
 	}
 }
