@@ -32,19 +32,19 @@ func backgroundColor(in term.File, out term.File) (color.Color, error) {
 // This function is intended for standalone Lip Gloss use only. If you're using
 // Bubble Tea, listen for tea.BackgroundColorMsg in your update function.
 func BackgroundColor(in term.File, out term.File) (bg color.Color, err error) {
-	if runtime.GOOS == "windows" {
+	if runtime.GOOS == "windows" { //nolint:nestif
 		// On Windows, when the input/output is redirected or piped, we need to
 		// open the console explicitly.
 		// See https://learn.microsoft.com/en-us/windows/console/getstdhandle#remarks
 		if !term.IsTerminal(in.Fd()) {
-			f, err := os.OpenFile("CONIN$", os.O_RDWR, 0o644)
+			f, err := os.OpenFile("CONIN$", os.O_RDWR, 0o644) //nolint:gosec
 			if err != nil {
 				return nil, fmt.Errorf("error opening CONIN$: %w", err)
 			}
 			in = f
 		}
 		if !term.IsTerminal(out.Fd()) {
-			f, err := os.OpenFile("CONOUT$", os.O_RDWR, 0o644)
+			f, err := os.OpenFile("CONOUT$", os.O_RDWR, 0o644) //nolint:gosec
 			if err != nil {
 				return nil, fmt.Errorf("error opening CONOUT$: %w", err)
 			}
