@@ -12,20 +12,12 @@ func btoi(b bool) int {
 	return 0
 }
 
-// max returns the greater of two integers.
-func max(a, b int) int { //nolint:predeclared
-	if a > b {
-		return a
+// bton converts a boolean to a specific integer, n if true, 0 if false.
+func bton(b bool, n int) int {
+	if b {
+		return n
 	}
-	return b
-}
-
-// min returns the smaller of two integers.
-func min(a, b int) int { //nolint:predeclared
-	if a < b {
-		return a
-	}
-	return b
+	return 0
 }
 
 // sum returns the sum of all integers in a slice.
@@ -49,16 +41,4 @@ func median(n []int) int {
 		return (n[h-1] + n[h]) / 2 //nolint:mnd
 	}
 	return n[len(n)/2]
-}
-
-// largest returns the largest element and it's index from a slice of integers.
-func largest(n []int) (int, int) { //nolint:unparam
-	var largest, index int
-	for i, e := range n {
-		if n[i] > n[index] {
-			largest = e
-			index = i
-		}
-	}
-	return index, largest
 }
